requirements.txt
Django
django-autoslug
django-bootstrap3
django-debug-toolbar
django-tastypie
